Unidad 2/Ejercicio 1/requirements.txt
uvicorn[standard]
python-dotenv
fastapi
fastapi[standard]
mysql.connector.python
requests
dotenv
pydantic
sqlmodel
